correct inventaire

+++ sql/llx_inventaire_app.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS `llx_inventaire_app` (
 `row_id` int(11) unsigned AUTO_INCREMENT NOT NULL,
   `k_inventaire_id` int(11) NOT NULL,
   `k_product_id` int(11) NOT NULL,
-  `pmp` float(6,6) NOT NULL,
+  `pmp` DOUBLE (24,8) NOT NULL,
   `reel` double DEFAULT NULL,
   `applied` int(1) NOT NULL DEFAULT '0',
   `date_modified` datetime DEFAULT NULL,
